traducao das struturas em go para ingles
