--- doc-center-api/models/arquivo.go
@@ -1,8 +1,0 @@
-package models
-
-type Arquivo struct {
-	Id        int64  `json:"id"`
-	Descricao string `json:"descricao"`
-	Arquivo   []byte `json:"arquivo"`
-	IdUsuario int64  `json:"usuario"`
-}
