+++ doc-center-api/models/dependent.go
@@ -1,12 +1,1 @@
 package models
-
-import "time"
-
-type Dependent struct {
-	Id            int64     `json:"id" gorm:"primaryKey"`
-	Name          string    `json:"name"`
-	BornDate      time.Time `json:"bornDate"`
-	ResponsableId int64     `json:"responsableId"`
-
-	User User `gorm:"foreignKey:ResponsableId"`
-}
