+++ doc-center-api/models/permission-type.go
@@ -1,6 +1,6 @@
 package models
 
-type PermissionType struct {
-	Id          int64  `json:"id" gorm:"primaryKey"`
-	Description string `json:"description"`
+type TipoPermissao struct {
+	Id        int64  `json:"id"`
+	Descricao string `json:"descricao"`
 }
