+++ doc-center-api/models/permission.go
@@ -1,18 +1,12 @@
 package models
 
-import (
-	"time"
-)
+import "time"
 
-type Permission struct {
-	Id        int64     `json:"id" gorm:"primaryKey"`
+type Permissao struct {
+	Id        int64     `json:"id"`
 	FileId    int64     `json:"file"`
-	OwnerId   int64     `json:"ownerUser"`
-	GuestId   int64     `json:"guestUser"`
+	Ownerid   int64     `json:"owner-user"`
+	GuestId   int64     `json:"guest-user"`
 	Period    time.Time `json:"period"`
 	Permanent bool      `json:"permanent"`
-
-	File      File `gorm:"foreignKey:FileId"`
-	OwnerUser User `gorm:"foreignKey:OwnerId"`
-	GuestUser User `gorm:"foreignKey:GuestId"`
 }
