--- doc-center-api/models/tipo-permissao.go
@@ -1,6 +1,0 @@
-package models
-
-type TipoPermissao struct {
-	Id        int64  `json:"id"`
-	Descricao string `json:"descricao"`
-}
