+++ doc-center-api/models/user.go
@@ -2,12 +2,12 @@ package models
 
 import "time"
 
-type User struct {
+type Usuario struct {
 	Id            int64     `json:"id" gorm:"primaryKey"`
 	Name          string    `json:"name"`
 	CPF           string    `json:"cpf" gorm:"check_cpf, pessoafisica = 1"`
 	BornDate      time.Time `json:"bornDate"`
-	CNPJ          string    `json:"cnpj" gorm:"check: check_copy, pessoafisica = 0"`
+	CNPJ          string    `json:"cnpj" gorm:"check: check_cnpj, pessoafisica = 0"`
 	NaturalPerson bool      `json:"naturalPerson"`
 	CompanyName   string    `json:"companyName"`
 	TradingName   string    `json:"tradingName"`
